pyann-model-executor/requirements.txt
Flask~=3.1
requests~=2.32
python-socketio[client]~=5.12
